+++ alacritty/src/input.rs
@@ -140,6 +140,15 @@ impl<T: EventListener> Execute<T> for Action {
                     ctx.clear_selection();
                 }
             },
+            #[cfg(not(any(target_os = "macos", windows)))]
+            Action::CopyPrimary => {
+                ctx.copy_selection(ClipboardType::Selection);
+
+                // Clear selection in vi mode for better user feedback
+                if ctx.terminal().mode().contains(TermMode::VI) {
+                    ctx.clear_selection();
+                }
+            },
             Action::Paste => {
                 let text = ctx.terminal_mut().clipboard().load(ClipboardType::Clipboard);
                 paste(ctx, &text);