+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Default Command+N keybinding for SpawnNewInstance on macOS
 - Vi mode for copying text and opening links
+- `CopyPrimary` action which copies into primary selection buffer on Linux/BSD
 
 ### Changed